pyproject.toml
[project]
name = "yolox-wrapper"
version = "0.1.0"
description = "YOLOX training, inference and deployment toolkit"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    # torch/torchvision は別途インストール (CUDA 版 / CPU 版)
    # 工場機 (CUDA 12.6):
    #   uv pip install --no-index --find-links ./wheels torch==2.6.0+cu126 torchvision==0.21.0+cu126
    # 開発機 (CPU):
    #   uv pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
    # YOLOX:
    #   uv pip install git+https://github.com/Megvii-BaseDetection/YOLOX.git
    #   uv pip install pycocotools-windows  # Windows の場合
    "numpy>=1.24,<2.0",
    "opencv-python>=4.8",
    "pillow>=10.0",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.9",
    "ty>=0.0.0a1",
]

[project.scripts]
yolox-gui = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/yolox_wrapper"]

# ---------------------------------------------------------------------------
# Ruff
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
src = ["src"]
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501"]

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
